// src/app/shared/models/club-search-result.ts
import { CustomKit } from './custom-kit';

export interface ClubSearchResult {
  clubId: string;
  name: string;
  rank: string;
  clubname: string;
  seasons: number;
  divGroupsWon: string;
  leaguesWon: string;
  divGroupsWon1: string;
  divGroupsWon2: string;
  divGroupsWon3: string;
  divGroupsWon4: string;
  cupsWon1: string;
  cupsWon2: string;
  cupsWon3: string;
  cupsWon4: string;
  cupsWon5: string;
  cupsElim1: string;
  cupsElim2: string;
  cupsElim3: string;
  cupsElim4: string;
  cupsElim5: string;
  promotions: string;
  holds: string;
  relegations: string;
  rankingPoints: string;
  curCompetition: string;
  prevDivision: string;
  prevGameDivision: string;
  bestDivision: number;
  bestPoints: string;
  curSeasonMov: string;
  recentResult0: string;
  recentResult1: string;
  recentResult2: string;
  recentResult3: string;
  recentResult4: string;
  recentResult5: string;
  recentResult6: string;
  recentResult7: string;
  recentResult8: string;
  recentResult9: string;
  recentOpponent0: string;
  recentOpponent1: string;
  recentOpponent2: string;
  recentOpponent3: string;
  recentOpponent4: string;
  recentOpponent5: string;
  recentOpponent6: string;
  recentOpponent7: string;
  recentOpponent8: string;
  recentOpponent9: string;
  recentScore0: string;
  recentScore1: string;
  recentScore2: string;
  recentScore3: string;
  recentScore4: string;
  recentScore5: string;
  recentScore6: string;
  recentScore7: string;
  recentScore8: string;
  recentScore9: string;
  wins: string;
  losses: string;
  ties: string;
  otl: string;
  prevSeasonWins: string;
  prevSeasonLosses: string;
  prevSeasonTies: string;
  prevSeasonOtl: string;
  goals: string;
  goalsAgainst: string;
  starLevel: string;
  totalCupsWon: number;
  cupsEntered: string;
  cupWinPercent: string;
  titlesWon: string;
  prevGameWonTitle: string;
  record: string;
  divsWon1: number;
  divsWon2: number;
  divsWon3: number;
  divsWon4: number;
  currentDivision: number;
  clubInfo: ClubInfo;
}

export interface ClubInfo {
  name: string;
  clubId: number;
  regionId: number;
  teamId: number;
  customKit: CustomKit;
}